Add support for disabling triggers
Added the ability to disable (as in, ignore) triggers that are connected to the MonsterShield, and to later re-enable them. This can only be done via the serial port.  Command to ignore all triggers (except manual triggers from button presses or serial interface) is "@I" (uppercase ) and to re-enable them the command is "@i" (lowercase i).  This is handy to make your MonsterShield ignore the triggers while you are trying to program it.