Python Web Server for MonsterShield
Simple Python web server interface that provides a web page that allows you to manually trigger any of the 15 sequences with a click of a button.  You can run this on a Raspberry Pi and have the MonsterShield / Arduino connected to one of the Raspberry Pi USB ports.  You will need to update the serial port to match the one you are using.  i.e., replace the '' string below:
ser = serial.Serial('/dev/ttyACM0',115200)

Windows example:
ser = serial.Serial('COM3',115200)

If you intend to use an SSL cert, you will also need to uncomment the 4 lines near the end of the code and provide your cert files.